+++ src/main/java/net/frozenblock/wilderwild/client/WWBlockRenderLayers.java
@@ -63,6 +63,7 @@ public final class WWBlockRenderLayers {
 		renderLayerRegistry.putBlock(WWBlocks.DATURA, RenderType.cutout());
 		renderLayerRegistry.putBlock(WWBlocks.CATTAIL, RenderType.cutout());
 		renderLayerRegistry.putBlock(WWBlocks.ALGAE, RenderType.cutout());
+		renderLayerRegistry.putBlock(WWBlocks.BARNACLES, RenderType.cutout());
 		renderLayerRegistry.putBlock(WWBlocks.MILKWEED, RenderType.cutout());
 		renderLayerRegistry.putBlock(WWBlocks.MARIGOLD, RenderType.cutout());
 		renderLayerRegistry.putBlock(WWBlocks.POTTED_MARIGOLD, RenderType.cutout());