+++ src/main/java/net/frozenblock/wilderwild/tag/WWBlockTags.java
@@ -158,6 +158,7 @@ public final class WWBlockTags {
 	public static final TagKey<Block> SMALL_COARSE_DIRT_PATH_REPLACEABLE = bind("feature/small_coarse_dirt_path_replaceable");
 	public static final TagKey<Block> PACKED_MUD_PATH_BADLANDS_REPLACEABLE = bind("feature/packed_mud_path_badlands_replaceable");
 	public static final TagKey<Block> POLLEN_FEATURE_PLACEABLE = bind("feature/pollen_feature_placeable");
+	public static final TagKey<Block> BARNACLES_FEATURE_PLACEABLE = bind("feature/barnacles_feature_placeable");
 	public static final TagKey<Block> TERMITE_DISK_REPLACEABLE = bind("feature/termite_disk_replaceable");
 	public static final TagKey<Block> TERMITE_DISK_BLOCKS = bind("feature/termite_disk_blocks");
 	public static final TagKey<Block> BLUE_NEMATOCYST_FEATURE_PLACEABLE = bind("feature/blue_nematocyst_feature_placeable");